Vard user4

+++ ACA.DeliverySystem.Data/Repository/OrderRepository.cs
@@ -31,13 +31,21 @@ namespace ACA.DeliverySystem.Data.Repository
 
         public async Task<OperationResult> Delete(int id, CancellationToken token)
         {
-            var order = await _context.Orders.SingleOrDefaultAsync(x => x.Id == id, token);
+            var order = await _context.Orders.Include(o => o.Items).SingleOrDefaultAsync(x => x.Id == id, token);
 
             if (order == null)
             {
                 return OperationResult.Error($"Order with id {id} not found.", ErrorType.NotFound);
             }
-
+            if (order.ProgressEnum == ProgressEnum.InProgress ||
+               order.ProgressEnum == ProgressEnum.Completed)
+            {
+                return OperationResult.Error($"You can't delete it. Order is {order.ProgressEnum}", ErrorType.BadRequest);
+            }
+            if (order.Items.Count != 0)
+            {
+                return OperationResult.Error($"Order have items in it. You can't delete it.", ErrorType.BadRequest);
+            }
             _context.Orders.Remove(order);
             return OperationResult.Ok();
         }
@@ -63,7 +71,7 @@ namespace ACA.DeliverySystem.Data.Repository
             }
             if (order.ProgressEnum != ProgressEnum.Created)
             {
-                return OperationResult.Error("You can't add item from order.", ErrorType.BadRequest);
+                return OperationResult.Error($"You can't add item. Order is {order.ProgressEnum}", ErrorType.BadRequest);
             }
             order.Items.Add(item);
             order.AmountToPay += item.Price;
@@ -89,7 +97,7 @@ namespace ACA.DeliverySystem.Data.Repository
             }
             if (order.ProgressEnum != ProgressEnum.Created)
             {
-                return OperationResult.Error("You can't remove item from order.", ErrorType.BadRequest);
+                return OperationResult.Error($"You can't remove item from order. Order is {order.ProgressEnum}", ErrorType.BadRequest);
             }
             order.Items.Remove(item);
             order.AmountToPay -= item.Price;
@@ -110,7 +118,7 @@ namespace ACA.DeliverySystem.Data.Repository
             }
             if (amount != amountToPay)
             {
-                return OperationResult.Error($"You must pay {amountToPay}.", ErrorType.BadRequest);
+                return OperationResult.Error($"You must pay {amountToPay.ToString("C")}.", ErrorType.BadRequest);
             }
             if (order.ProgressEnum != ProgressEnum.Created)
             {
@@ -131,7 +139,7 @@ namespace ACA.DeliverySystem.Data.Repository
             }
             if (order.ProgressEnum != ProgressEnum.InProgress)
             {
-                return OperationResult.Error("Order must be in progress.", ErrorType.BadRequest);
+                return OperationResult.Error("Order must be in progress to marke it as completed.", ErrorType.BadRequest);
             }
             order.ProgressEnum = ProgressEnum.Completed;
             return OperationResult.Ok();
@@ -152,6 +160,10 @@ namespace ACA.DeliverySystem.Data.Repository
             if (order.ProgressEnum == ProgressEnum.Canceled)
             {
                 return OperationResult.Error("It's already canceled.", ErrorType.BadRequest);
+            }
+            if (order.ProgressEnum == ProgressEnum.InProgress)
+            {
+                return OperationResult.Error("You can't cancel order.It is in progress.", ErrorType.BadRequest);
             }
             order.ProgressEnum = ProgressEnum.Canceled;
             return OperationResult.Ok();
